convert / to server-side initial data fetch
page.tsx is now an async server component that queries top_100_delegates,
last event timestamp, and votable supply in parallel, then hands them to
HomeClient as props. kills the post-hydration fetch and empty-table flash
on first paint. all client interactivity preserved (search, sort,
pagination, modal, csv export, proposer stats, vote history).
